% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/menv.R
\name{ff.bind_path}
\alias{ff.bind_path}
\title{Bind an absolute path for a kind of resources.}
\usage{
ff.bind_path(type, sub_path)
}
\arguments{
\item{type}{the type of resource.}

\item{sub_path}{the suffix to concatenate to the absolute path to get the absolute path of the kind of resource.}
}
\description{
Bind an absolute path for a kind of resources.
}
\examples{
ff.bind_sub_path(type = "data",sub_path = "dataset")
}

