% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/menv.R
\name{ff.bindPath}
\alias{ff.bindPath}
\title{Bind an absolute path for a kind of resources.}
\usage{
ff.bindPath(type, sub_path)
}
\arguments{
\item{type}{the type of resource.}

\item{sub_path}{the suffix to concatenate to the absolute path to get the absolute path of the kind of resource.}
}
\description{
Bind an absolute path for a kind of resources.
}
\examples{
ff.setBasePath(getwd())
if(! dir.exists("mydata") ) dir.create('mydata')
ff.bindPath(type = "data",sub_path = "mydata")
}

