% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/featureFilter.R
\name{ff.corrFilter}
\alias{ff.corrFilter}
\title{Filter a \code{data.frame} of numeric according to a given threshold of correlation}
\usage{
ff.corrFilter(Xtrain, Xtest, y, abs_th = NULL, rel_th = 1,
  method = "pearson")
}
\arguments{
\item{Xtrain}{a train set \code{data.frame} of \code{numeric}}

\item{Xtest}{a test set \code{data.frame} of \code{numeric}}

\item{y}{the output variable (as numeric vector)}

\item{abs_th}{an absolute threshold (= number of data frame columns)}

\item{rel_th}{a relative threshold (= percentage of data frame columns)}

\item{method}{a character string indicating which correlation method is to be used for the test. One of "pearson", "kendall", or "spearman".}
}
\value{
a \code{list} of filtered train set and test set with correlation test results
}
\description{
Filter a \code{data.frame} of numeric according to a given threshold of correlation
}
\examples{
Xtrain <- data.frame( a = rep(1:3 , each = 2), b = c(4:1,6,6), c = rep(1,6))
Xtest <-  Xtrain + runif(nrow(Xtrain))
y = 1:6
l = ff.corrFilter(Xtrain=Xtrain,Xtest=Xtest,y=y,rel_th=0.5)
Xtrain.filtered = l$Xtrain
Xtest.filtered =l$Xtest
}

