% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/featureFilter.R
\name{ff.featureFilter}
\alias{ff.featureFilter}
\title{Filter predictors according to specified criteria.}
\usage{
ff.featureFilter(traindata, testdata, y = NULL,
  removeOnlyZeroVariacePredictors = FALSE,
  performVarianceAnalysisOnTrainSetOnly = TRUE, correlationThreshold = NA,
  removePredictorsMakingIllConditionedSquareMatrix = TRUE,
  removeHighCorrelatedPredictors = TRUE, featureScaling = TRUE,
  verbose = TRUE)
}
\arguments{
\item{traindata}{the train set}

\item{testdata}{the test set}

\item{y}{the response variable}

\item{removeOnlyZeroVariacePredictors}{\code{TRUE} to remove only zero variance predictors}

\item{performVarianceAnalysisOnTrainSetOnly}{\code{TRUE} to perform the variance analysis on the train set only}

\item{correlationThreshold}{a correlation threshold above which keeping predictors
(considered only if \code{removeOnlyZeroVariacePredictors} is \code{FALSE}).}

\item{removePredictorsMakingIllConditionedSquareMatrix}{\code{TRUE} to predictors making ill conditioned square matrices}

\item{removeHighCorrelatedPredictors}{\code{TRUE} to remove high correlared predictors}

\item{featureScaling}{\code{TRUE} to perform feature scaling}

\item{verbose}{\code{TRUE} to set verbose mode}
}
\description{
Filter predictors according to specified criteria.
}
\examples{
Xtrain <- data.frame( a = rep(1:3 , each = 2), b = c(4:1,6,6), c = rep(1,6))
Xtest <-  Xtrain + runif(nrow(Xtrain))
l = ff.featureFilter (traindata = Xtrain,
                       testdata = Xtest,
                       removeOnlyZeroVariacePredictors=TRUE)
Xtrain = l$traindata
Xtest = l$testdata
}

