% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/menv.R
\name{ff.get_bindings}
\alias{ff.get_bindings}
\title{Get the list of bindings (type of resource , absolutepath)}
\usage{
ff.get_bindings()
}
\description{
Get the list of bindings (type of resource , absolutepath)
}
\examples{
ff.get_bindings()
}

