% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/menv.R
\name{ff.get_path}
\alias{ff.get_path}
\title{Get the absolute path for a kind of resources.}
\usage{
ff.get_path(type = "base")
}
\arguments{
\item{type}{the type of resource.}
}
\description{
Get the absolute path for a kind of resources.
}
\examples{
ff.get_path() ## equivalent to ff.get_path = function (type="base") gets the base path
}

