% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/menv.R
\name{ff.get_path_bindings}
\alias{ff.get_path_bindings}
\title{Get the list of bindings (type of resource , absolutepath)}
\usage{
ff.get_path_bindings()
}
\description{
Get the list of bindings (type of resource , absolutepath)
}
\examples{
ff.set_base_path(getwd())
if(! dir.exists("mydata") ) dir.create('mydata')
ff.bind_path(type = "data",sub_path = "mydata")
ff.get_path_bindings()
}

