% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastRegression.R
\name{ff.getBestBlenderPerformance}
\alias{ff.getBestBlenderPerformance}
\title{Helper function that given a blender object returns the best optimization method.}
\usage{
ff.getBestBlenderPerformance(blender)
}
\arguments{
\item{blender}{a blender object}
}
\value{
a \code{numeric} of best score and as object name the best performant method name.
}
\description{
Helper function that given a blender object returns the best optimization method.
}
\seealso{
\code{\link[fastfurious]{ff.blend}} for examples.
}

