% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastRegression.R
\name{ff.getBestBlenderTune}
\alias{ff.getBestBlenderTune}
\title{Helper function that given a blender object returns the best tuning parameters found by the blender.}
\usage{
ff.getBestBlenderTune(blender, truncate = TRUE)
}
\arguments{
\item{blender}{a blender object}

\item{truncate}{\code{TRUE} to cut at the first tuning best configuration in case there are more than one optimal tuning configurations.}
}
\value{
a \code{data.frame} of the best tuning parameters.
}
\description{
Helper function that given a blender object returns the best tuning parameters found by the blender.
}
\seealso{
\code{\link[fastfurious]{ff.blend}} for examples.
}

