% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/menv.R
\name{ff.getMaxCuncurrentThreads}
\alias{ff.getMaxCuncurrentThreads}
\title{Get the max number of cuncurrent threads.}
\usage{
ff.getMaxCuncurrentThreads()
}
\description{
Get the max number of cuncurrent threads.
}
\examples{
ff.getMaxCuncurrentThreads()
}

