% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/menv.R
\name{ff.getPathBindings}
\alias{ff.getPathBindings}
\title{Get the list of bindings, i.e. (type resource,absolute path) pairs as a list}
\usage{
ff.getPathBindings()
}
\value{
the list of bindings
}
\description{
Get the list of bindings, i.e. (type resource,absolute path) pairs as a list
}
\examples{
ff.setBasePath(getwd())
if(! dir.exists("mydata") ) dir.create('mydata')
ff.bindPath(type = "data",sub_path = "mydata")
ff.getPathBindings()
}

