% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastUnsupervised.R
\name{ff.kmeans}
\alias{ff.kmeans}
\title{An useful wrapper of \code{\link[stats]{kmeans}} performing k-means clustering on the given trainset / testset (\code{Xtrain} / \code{Xtest}) and
assuming a number of cluster from 1:\code{max_centers}. The best number of cluster is computed so that the variation in
the within group sum of squares between two subsequent number of clusters is maximized in absolute value.}
\usage{
ff.kmeans(Xtrain, Xtest, max_centers = 10, nstart = 5, iter.max = 10,
  doPlot = FALSE, verbose = FALSE)
}
\arguments{
\item{Xtrain}{the encoded \code{data.frame} of train data. Must be a \code{data.frame} of \code{numeric}}

\item{Xtest}{the encoded \code{data.frame} of train data. Must be a \code{data.frame} of \code{numeric}}

\item{max_centers}{the max number of clusters to be evaluated}

\item{nstart}{how many random sets should be chosen? Such a parameter is passed to \code{\link[stats]{kmeans}}.}

\item{iter.max}{the maximum number of iterations allowed. Such a parameter is passed to \code{\link[stats]{kmeans}}.}

\item{doPlot}{a logical value indicating whether plotting.}

\item{verbose}{a logical value indicating whether verbose mode should be enabled.}

\item{varThreshold}{a threshold indicating the proportion of variance that should be explained. Must be a numeric between 0 and 1.}
}
\value{
a list whose components are the max number of clusters evaluated (\code{max_centers}), the best number of clusters
 (\code{best_n_cluters}), assuming \code{best_n_cluters} as number of clusters a vector of integers (from 1:\code{best_n_cluters})
 indicating the cluster to which each point is allocated in the train set (\code{K.train}) and test set (\code{K.test}).
}
\description{
An useful wrapper of \code{\link[stats]{kmeans}} performing k-means clustering on the given trainset / testset (\code{Xtrain} / \code{Xtest}) and
assuming a number of cluster from 1:\code{max_centers}. The best number of cluster is computed so that the variation in
the within group sum of squares between two subsequent number of clusters is maximized in absolute value.
}
\examples{
## data
Xtrain <- data.frame(a = rep(1:10 , each = 2), b = 20:1,
                     c = rep(as.Date(c("2007-06-22", "2004-02-13")),10) , d = 20:1)
Xtest <- data.frame(a = rep(2:11 , each = 2), b = 1:20,
                    c = rep(as.Date(c("2007-03-01", "2004-05-23")),10) , d = 1:20)

## encode data sets
l = ff.makeFeatureSet(Xtrain,Xtest,c("C","N","D","N"))
Xtrain = l$traindata
Xtest = l$testdata

ffKmeans = ff.kmeans(Xtrain = Xtrain , Xtest = Xtest ,
                     max_centers = 10 , verbose = TRUE)

best_n_cluters = ffKmeans$best_n_cluters

K.train = ffKmeans$K.train
K.test = ffKmeans$K.test
}

