% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastUnsupervised.R
\name{ff.pca}
\alias{ff.pca}
\title{An useful wrapper of \code{\link[stats]{prcomp}} performing a principal components analysis on the given trainset / testset (\code{Xtrain} / \code{Xtest}).}
\usage{
ff.pca(Xtrain, Xtest, center = TRUE, scale. = FALSE,
  removeZeroVarPredictors = TRUE, varThreshold = 0.95, doPlot = TRUE,
  verbose = FALSE)
}
\arguments{
\item{Xtrain}{the encoded \code{data.frame} of train data. Must be a \code{data.frame} of \code{numeric}}

\item{Xtest}{the encoded \code{data.frame} of train data. Must be a \code{data.frame} of \code{numeric}}

\item{center}{a logical value indicating whether the variables should be shifted to be zero centered.
Alternately, a vector of length equal the number of columns of data can be supplied. The value is passed to \code{scale}.}

\item{scale.}{a logical value indicating whether the variables should be scaled to have unit variance before the analysis takes place.
The default is \code{FALSE} for consistency with S, but in general scaling is advisable.
Alternatively, a vector of length equal the number of columns of data can be supplied. The value is passed to \code{scale}.}

\item{removeZeroVarPredictors}{a logical value indicating whether removing zero variance predictors before calling \code{\link[stats]{prcomp}}
preventing errors due to the fact that the latter cannot rescale a constant/zero column to unit variance.}

\item{varThreshold}{a threshold indicating the proportion of variance that should be explained. Must be a numeric between 0 and 1.}

\item{doPlot}{a logical value indicating whether plotting the proportion of variance explained vs. principal components.}

\item{verbose}{a logical value indicating whether verbose mode should be enabled.}
}
\value{
a list whose components are the number of principal components (\code{numComp}), the number of principal components to hold
so that the proportion of variance
explained by each subsequent principal component drops off as an elbow in the screen plot (\code{numComp.elbow}), the number of principal
components explaining a given (specified by the \code{varThreshold} input parameter) proportion of variance (\code{numComp.threshold}),
the threshold indicating the proportion of variance that should be explained (\code{varThreshold}),
the cumulative sum of proportion of variance explained by each principal component (\code{cumVar}), the proportion of variance
explained by each principal component (\code{var}), the principal components for train and test set (\code{PC.train} and \code{PC.test})
}
\description{
An useful wrapper of \code{\link[stats]{prcomp}} performing a principal components analysis on the given trainset / testset (\code{Xtrain} / \code{Xtest}).
}
\examples{
## data
Xtrain <- data.frame( a = rep(1:10 , each = 2), b = 20:1, c = rep(as.Date(c("2007-06-22", "2004-02-13")),10) , d = 20:1)
Xtest <- data.frame( a = rep(2:11 , each = 2), b = 1:20, c = rep(as.Date(c("2007-03-01", "2004-05-23")),10) , d = 1:20)

## encode data sets
l = ff.makeFeatureSet(Xtrain,Xtest,c("C","N","D","N"))
Xtrain = l$traindata
Xtest = l$testdata

ffPCA = ff.pca(Xtrain = Xtrain , Xtest = Xtest , center = TRUE , scale. = TRUE , removeZeroVarPredictors = TRUE ,
               varThreshold = 0.95 , doPlot = FALSE , verbose = TRUE)

numComp <- ffPCA$numComp
numComp.elbow <- ffPCA$numComp.elbow
numComp.threshold <- ffPCA$numComp.threshold

Xtrain_95Var = ffPCA$PC.train[1:numComp.threshold,,drop=FALSE]
Xtest_95Var = ffPCA$PC.test[1:numComp.threshold,,drop=FALSE]
}

