% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastRegression.R
\name{ff.plotPerformance.reg}
\alias{ff.plotPerformance.reg}
\title{Plot predicted values vs. observed / residual values.}
\usage{
ff.plotPerformance.reg(observed, predicted, main = NULL)
}
\arguments{
\item{observed}{the observed output variables (numeric vector).}

\item{predicted}{the predicted values (numeric vector).}

\item{main}{a string as a title for the plot}
}
\description{
Plot predicted values vs. observed / residual values.
}
\examples{
obs = 1:10
preds = obs + runif(length(obs))
ff.plotPerformance.reg(observed = obs , predicted = preds, main="Predicted vs. observed/residual")
}

