% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/featureFilter.R
\name{ff.poly}
\alias{ff.poly}
\title{Make polynomial terms of a \code{data.frame}}
\usage{
ff.poly(x, n, direction = 0)
}
\arguments{
\item{x}{a \code{data.frame} of \code{numeric}}

\item{n}{the polynomial degree}

\item{direction}{if set to \code{0} returns the terms \code{x^(1/n),x^(1/(n-1)),...,x,x^2,...,x^n}.
If set to \code{-1} returns the terms \code{x^(1/n),x^(1/(n-1)),...,x}.
If set to \code{1} returns the terms \code{x,x^2,...,x^n}.}
}
\value{
the \code{data.frame} with the specified polynomial terms
}
\description{
Make polynomial terms of a \code{data.frame}
}
\examples{
Xtrain <- data.frame( a = rep(1:3 , each = 2), b = c(4:1,6,6), c = rep(1,6))
Xtest <-  Xtrain + runif(nrow(Xtrain))
data = rbind(Xtrain,Xtest)
data.poly = ff.poly(x=data,n=3)
Xtrain.poly = data.poly[1:nrow(Xtrain),]
Xtest.poly = data.poly[(nrow(Xtrain)+1):nrow(data),]
}

