% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/menv.R
\name{ff.set_base_path}
\alias{ff.set_base_path}
\title{Set base path}
\usage{
ff.set_base_path(path)
}
\arguments{
\item{path}{the absolute path.}
}
\description{
Set base path
}
\examples{
ff.set_base_path('/Users/gg/root')
}

