% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/menv.R
\name{ff.setBasePath}
\alias{ff.setBasePath}
\title{Set base path}
\usage{
ff.setBasePath(path)
}
\arguments{
\item{path}{the absolute path.}
}
\description{
Set base path
}
\examples{
ff.setBasePath('./')
}

