% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/menv.R
\name{ff.setMaxCuncurrentThreads}
\alias{ff.setMaxCuncurrentThreads}
\title{Set the max number of cuncurrent threads.}
\usage{
ff.setMaxCuncurrentThreads(nThreads = 2)
}
\arguments{
\item{nThreads}{max number of cuncurrent threads.}
}
\description{
Set the max number of cuncurrent threads.
}
\examples{
ff.setMaxCuncurrentThreads(4)
}

