% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastRegression.R
\name{ff.summaryBlender}
\alias{ff.summaryBlender}
\title{Helper function that given a blender object returns a \code{numeric} vector of performances (one for each optimization method).}
\usage{
ff.summaryBlender(blender)
}
\arguments{
\item{blender}{a blender object}
}
\value{
a \code{numeric} vector of performances (one for each optimization method)
}
\description{
Helper function that given a blender object returns a \code{numeric} vector of performances (one for each optimization method).
}
\seealso{
\code{\link[fastfurious]{ff.blend}} for examples.
}

