% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastRegression.R
\name{ff.verifyBlender}
\alias{ff.verifyBlender}
\title{Helper function that given a blender object replicates the execution in order to verify performances.}
\usage{
ff.verifyBlender(blender, Xtrain, y, seed = NULL, controlObject,
  caretModelname)
}
\arguments{
\item{blender}{a blender object}

\item{Xtrain}{the train set}

\item{y}{the output variable as numeric vector}

\item{seed}{the seed used by the blender, if applicable. If the blender used one, it is necessary for replicating blender performances.}

\item{controlObject}{a list of values that define how this function acts. Must be a caret \code{trainControl} object. It must be the same used by the blender.}

\item{caretModelname}{a string specifying which model to use. Possible values are \code{'lm'}, \code{'bayesglm'}, 
\code{'glm'}, \code{'glmStepAIC'}, \code{'rlm'}, \code{'knn'}, \code{'pls'}, \code{'ridge'}, \code{'enet'}, 
\code{'svmRadial'}, \code{'treebag'}, \code{'gbm'}, \code{'rf'}, \code{'cubist'}, \code{'avNNet'}, 
\code{'xgbTreeGTJ'}, \code{'xgbTree'}. It must be the same model name used by the blender.}
}
\value{
a \code{numeric} as difference in performance between blender and replicated execution.
}
\description{
Helper function that given a blender object replicates the execution in order to verify performances.
}
\seealso{
\code{\link[fastfurious]{ff.blend}} for examples.
}

