% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastRegression.R
\name{RMSE.xgb}
\alias{RMSE.xgb}
\title{Root mean square error}
\usage{
RMSE.xgb(preds, dtrain)
}
\arguments{
\item{preds}{the predicted values (numeric vector).}

\item{dtrain}{the xgboost train set object.}
}
\value{
a list of metric label / values
}
\description{
Root mean square error
}

