% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastRegression.R
\name{RMSLE.xgb}
\alias{RMSLE.xgb}
\title{Root mean square logistic error}
\usage{
RMSLE.xgb(preds, dtrain, th_err = 1.5)
}
\arguments{
\item{preds}{the predicted values (numeric vector).}

\item{dtrain}{the xgboost train set object.}

\item{th_err}{a threshold in case predictions are negative.}
}
\value{
a list of metric label / values
}
\description{
Root mean square logistic error
}

