% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastRegression.R
\name{RMSPE}
\alias{RMSPE}
\title{Root mean square percentage error}
\usage{
RMSPE(preds, dtrain)
}
\arguments{
\item{preds}{the predicted values (numeric vector).}

\item{dtrain}{the xgboost train set object.}
}
\value{
a list of metric label / values
}
\description{
Root mean square percentage error
}

