% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menv.R
\name{ff.getPath}
\alias{ff.getPath}
\title{Get the absolute path for a kind of resources.}
\usage{
ff.getPath(type = "base")
}
\arguments{
\item{type}{the type of resource.}
}
\value{
the absolute path for a kind of resources (as character)
}
\description{
Get the absolute path for a kind of resources.
}
\examples{
ff.setBasePath('./')
ff.getPath() ## equivalent to ff.getPath(type="base") 
}

